// src/app/model/calculateWorkedHoursModel.ts
export interface CalculateHours {

    baseSalary: number;
    hourlyPrice: number;
    normalWorkingHours: number;
    normalOvertimeWorked: number;
    workingNightHours: number;
    overnigthHoursWorked: number;
    sundayWorkingHours: number;
    sundayOvertimeWork: number;
    normalHours: number;
    nightHours: number;
    sundayHours: number;
    normalExtraHours: number;
    nightlyOvertime: number;
    sundayExtraHours: number;
    finalSalary: number;

}